File separation, bug fixes, and preparations for future
